api/speciesnet/requirements.txt
# Server dependencies
fastapi>=0.68.0
uvicorn>=0.15.0
requests>=2.28.0
litserve>=0.1.0
numpy>=1.21.0
Pillow>=9.0.0
absl-py>=1.0.0
torch>=2.0.0
torchvision>=0.15.0
cloudpathlib>=0.15.0
tqdm>=4.65.0
